requirements.txt
##############################################
# FramePack pinned requirements (Windows/Linux)
# Python 3.10 / CUDA 12.6 (recommended)
##############################################

# Prefer installing PyTorch with the official index for CUDA 12.6.
# If your environment already has PyTorch, you can comment out the next 4 lines.
--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0

accelerate==1.6.0
diffusers==0.33.1
transformers==4.46.2
gradio==5.23.0
sentencepiece==0.2.0
pillow==11.1.0
av==12.1.0
numpy==1.26.2
scipy==1.12.0
requests==2.31.0
torchsde==0.2.6

einops
opencv-contrib-python
safetensors

timm
kornia

# Optional accelerators (install only if your CUDA/toolchain matches)
# xformers==0.0.27.post2
# flash-attn==2.5.9
# sageattention==1.0.6

# Note: FFmpeg binary is required and must be installed separately
# Windows: https://ffmpeg.org/download.html
# Linux: sudo apt-get install -y ffmpeg
